refactor(ha): split Patroni rollout model
